talks/17_fastapi/requirements.txt
fastapi
uvicorn[standard]
pyyaml>=5.3.1
rich